finalise tests

+++ database/migrations/2024_04_22_134624_update_users_table.php
@@ -12,11 +12,13 @@ return new class extends Migration {
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->integer('purchasesNumber')->default('0');
-            $table->integer('idLastUpline')->nullable();
-            $table->integer('idReservedUpline')->nullable();
-        });
+        if (Schema::hasTable('users')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->integer('purchasesNumber')->default('0');
+                $table->integer('idLastUpline')->nullable();
+                $table->integer('idReservedUpline')->nullable();
+            });
+        }
     }
 
     /**
@@ -26,10 +28,12 @@ return new class extends Migration {
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('purchasesNumber');
-            $table->dropColumn('idLastUpline');
-            $table->dropColumn('idReservedUpline');
-        });
+        if (Schema::hasTable('users')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('purchasesNumber');
+                $table->dropColumn('idLastUpline');
+                $table->dropColumn('idReservedUpline');
+            });
+        }
     }
 };
